requirements.txt
python-dotenv>=1.0.1
openai>=1.0.0
python-telegram-bot==20.8
APScheduler>=3.10.4
